Fix pluginpref version API
Hexchat was trying to be smart and auto-converted things that "looked
like a number" into a number, but a string like "0.3.0" is not a number.
What hexchat stored instead was just "0" (as an integer, not a string).

We work against that by prefixing our "number-like" version string with
a "v" so that hexchat doesn't start getting ideas.